backend/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
python-dotenv
fastapi
uvicorn[standard]